requirements.txt
accelerate==0.31.0
datasets==3.0.1
decord==0.6.0
diffusers==0.30.2
einops==0.8.0
gradio==5.13.1
imageio==2.35.1
imageio-ffmpeg==0.5.1
pillow==10.4.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
transformers==4.41.1
deepspeed==0.15.1